s/processd: more work on the new PE loader
