+++ services/processd/pe/helpers.c
@@ -20,82 +20,14 @@
  *      and implemented as a part of libds to share between services and kernel
  */
 
-#define __TRACE
-
 #include <ddk/memory.h>
-#include <ddk/utils.h>
-#include <ds/mstring.h>
 #include <internal/_syscalls.h>
-#include <os/services/file.h>
-#include <os/memory.h>
 #include "pe.h"
-#include "process.h"
-#include <stdlib.h>
 #include <stdio.h>
-#include <string.h>
 #include <time.h>
 
 static SystemDescriptor_t g_systemInformation = { 0 };
 static uintptr_t          g_systemBaseAddress = 0;
-
-oserr_t
-PELoadImage(
-        _In_  mstring_t* fullPath,
-        _Out_ void**     bufferOut,
-        _Out_ size_t*    lengthOut)
-{
-    FILE*   file;
-    long    fileSize;
-    void*   fileBuffer;
-    size_t  bytesRead;
-    oserr_t osStatus = OS_EOK;
-    char*   pathu8;
-
-    pathu8 = mstr_u8(fullPath);
-    ENTRY("LoadFile %s", pathu8);
-
-    // special case:
-    // load from ramdisk
-    if (mstr_find_u8(fullPath, "/initfs/", 0) != -1) {
-        return PmBootstrapFindRamdiskFile(fullPath, bufferOut, lengthOut);
-    }
-
-    file = fopen(pathu8, "rb");
-    free(pathu8);
-    if (!file) {
-        ERROR("LoadFile fopen failed: %i", errno);
-        osStatus = OS_ENOENT;
-        goto exit;
-    }
-
-    fseek(file, 0, SEEK_END);
-    fileSize = ftell(file);
-    rewind(file);
-
-    TRACE("[load_file] size %" PRIuIN, fileSize);
-    fileBuffer = malloc(fileSize);
-    if (!fileBuffer) {
-        ERROR("LoadFile null");
-        fclose(file);
-        osStatus = OS_EOOM;
-        goto exit;
-    }
-
-    bytesRead = fread(fileBuffer, 1, fileSize, file);
-    fclose(file);
-
-    TRACE("LoadFile read %" PRIuIN " bytes from file", bytesRead);
-    if (bytesRead != fileSize) {
-        osStatus = OS_EINCOMPLETE;
-    }
-
-    *bufferOut = fileBuffer;
-    *lengthOut = fileSize;
-
-exit:
-    EXIT("LoadFile");
-    return osStatus;
-}
 
 uintptr_t
 PECurrentPageSize(void)
