+++ services/processd/pe/resolver.c
@@ -18,88 +18,23 @@
 #include <ddk/utils.h>
 #include <ds/mstring.h>
 #include <os/services/file.h>
+#include <os/services/path.h>
 #include <stdlib.h>
-
-static inline oserr_t
-__TestFilePath(
-        _In_ mstring_t* path)
-{
-    OsFileDescriptor_t fileDescriptor;
-    char*              pathu8;
-    oserr_t            osStatus;
-
-    pathu8 = mstr_u8(path);
-    if (pathu8 == NULL) {
-        return OS_EOOM;
-    }
-
-    osStatus = GetFileInformationFromPath(
-            pathu8,
-            1,
-            &fileDescriptor
-    );
-    free(pathu8);
-    return osStatus;
-}
-
-static oserr_t
-__GuessBasePath(
-        _In_  uuid_t      processHandle,
-        _In_  mstring_t*  path,
-        _Out_ mstring_t** fullPathOut)
-{
-    // Check the working directory, if it fails iterate the environment defaults
-    Process_t* process = PmGetProcessByHandle(processHandle);
-    mstring_t* result;
-    int        isApp;
-    int        isDll;
-
-    TRACE("__GuessBasePath(process=%u, path=%ms)",
-          processHandle, path);
-
-    // Start by testing against the loaders current working directory,
-    // however this won't work for the base process
-    if (process != NULL) {
-        result = mstr_fmt("%ms/%ms", process->working_directory, path);
-        if (__TestFilePath(result) == OS_EOK) {
-            *fullPathOut = result;
-            return OS_EOK;
-        }
-        mstr_delete(result);
-    }
-
-    // At this point we have to run through all PATH values
-    // Look at the type of file we are trying to load. .app? .dll?
-    // for other types its most likely resource load
-    isApp = mstr_find_u8(path, ".run", 0);
-    isDll = mstr_find_u8(path, ".dll", 0);
-    if (isApp != -1 || isDll != -1) {
-        result = mstr_fmt("/shared/bin/%ms", path);
-    } else {
-        result = mstr_fmt("/system/bin/%ms", path);
-    }
-
-    if (__TestFilePath(result) == OS_EOK) {
-        *fullPathOut = result;
-        return OS_EOK;
-    }
-    else {
-        mstr_delete(result);
-        return OS_EUNKNOWN;
-    }
-}
+#include <string.h>
+#include "process.h"
+#include "private.h"
 
 static mstring_t*
 __TestRamdiskPath(
         _In_ const char* basePath,
-        _In_ mstring_t*  path)
+        _In_ const char* path)
 {
     oserr_t    osStatus;
     mstring_t* temporaryResult;
     TRACE("__TestRamdiskPath(basePath=%s, path=%ms)", basePath, path);
 
     // create the full path for the ramdisk
-    temporaryResult = mstr_fmt("%s/%ms", basePath, path);
+    temporaryResult = mstr_fmt("%s/%s", basePath, path);
     if (temporaryResult == NULL) {
         return NULL;
     }
@@ -115,59 +50,61 @@ __TestRamdiskPath(
 
 static oserr_t
 __ResolveRelativePath(
-        _In_  uuid_t      processId,
-        _In_  mstring_t*  parentPath,
-        _In_  mstring_t*  path,
-        _Out_ mstring_t** fullPathOut)
+        _In_  struct PEImageLoadContext* loadContext,
+        _In_  const char*                path,
+        _Out_ mstring_t**                fullPathOut)
 {
-    oserr_t    osStatus;
-    mstring_t* temporaryResult = path;
+    oserr_t            oserr;
+    OsFileDescriptor_t fileDescriptor;
     TRACE("__ResolveRelativePath(processId=%u, parentPath=%ms, path=%ms)",
           processId, parentPath, path);
 
-    // Let's test against parent being loaded through the ramdisk
-    if (parentPath && mstr_find_u8(parentPath, "/initfs/", 0) != -1) {
-        // create the full path for the ramdisk
-        temporaryResult = __TestRamdiskPath("/initfs/bin", path);
-        if (!temporaryResult) {
-            // sometimes additional modules will be loaded (i.e fs modules)
-            temporaryResult = __TestRamdiskPath("/initfs/modules", path);
+    for (char* i = strtok(loadContext->Paths, ";"); i; i = strtok(NULL, ";")) {
+        char* combined = OSPathJoin(i, path);
+        if (strstr(combined, "/initfs/") != NULL) {
+            mstring_t* result = __TestRamdiskPath(i, path);
+            if (result) {
+                *fullPathOut = mstr_new_u8(combined);
+                free(combined);
+                break;
+            }
+        } else {
+            oserr = GetFileInformationFromPath(
+                    combined,
+                    1,
+                    &fileDescriptor
+            );
+            if (oserr == OS_EOK) {
+                *fullPathOut = mstr_new_u8(combined);
+                free(combined);
+                break;
+            }
         }
-
-        if (temporaryResult) {
-            *fullPathOut = temporaryResult;
-            return OS_EOK;
-        }
-
-        // restore temporaryResult
-        temporaryResult = path;
+        free(combined);
     }
-
-    osStatus = __GuessBasePath(processId, path, &temporaryResult);
-    TRACE("__ResolveRelativePath basePath=%ms", temporaryResult);
-
-    // If we already deduced an absolute path skip the canonicalizing moment
-    if (osStatus == OS_EOK && mstr_at(temporaryResult, 0) == U'/') {
-        *fullPathOut = temporaryResult;
-        return osStatus;
-    }
-    return OS_ENOENT;
+    return OS_EOK;
 }
 
 oserr_t
-PeImplResolveFilePath(
-        _In_  uuid_t      processId,
-        _In_  mstring_t*  parentPath,
-        _In_  mstring_t*  path,
-        _Out_ mstring_t** fullPathOut)
+PEResolvePath(
+        _In_  struct PEImageLoadContext* loadContext,
+        _In_  mstring_t*                 path,
+        _Out_ mstring_t**                fullPathOut)
 {
     oserr_t oserr = OS_EOK;
+    char*   cpath;
     ENTRY("ResolveFilePath(processId=%u, path=%ms)", processId, path);
 
     if (mstr_at(path, 0) != U'/') {
+        cpath = mstr_u8(path);
+        if (cpath == NULL) {
+            return OS_EOOM;
+        }
+
         // If we don't even have an environmental identifier present, we
         // have to get creative and guess away
-        oserr = __ResolveRelativePath(processId, parentPath, path, fullPathOut);
+        oserr = __ResolveRelativePath(loadContext, cpath, fullPathOut);
+        free(cpath);
     } else {
         // Assume absolute path
         *fullPathOut = mstr_clone(path);
